Move and enhance photo info box
