+++ lib/admin/photos/photos.module.css
@@ -38,7 +38,3 @@
     min-width: 250px;
     height: 30px;
 }
-
-.Error {
-    color: red;
-}
